pyproject.toml
[tool.poetry]
name = "fish-diffusion"
version = "0.1.0"
description = ""
authors = ["Lengyue <lengyue@lengyue.me>"]
license = "Apache"

packages = [{ include = "fish_diffusion" }]

[tool.poetry.dependencies]
python = "^3.10"
praat-parselmouth = "^0.4.3"
soundfile = "^0.11.0"
librosa = "^0.9.1"
pytorch-lightning = "^1.8.6"
numba = "^0.56.4"
fish-audio-preprocess = "^0.1.4"
wandb = "^0.13.9"
transformers = "^4.25.1"
torchcrepe = "^0.0.17"
mmengine = "^0.4.0"
loguru = "^0.6.0"
click = "^8.1.3"
tensorboard = "^2.11.2"
openai-whisper = "^20230124"
pypinyin = "^0.48.0"
TextGrid = "^1.5"
pyworld = "^0.3.2"
pykakasi = "^2.2.1"

[tool.poetry.dev-dependencies]
isort = "^5.11.4"
black = "^22.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
extend_skip = ["dataset", "logs"]
